+++ package.json
@@ -5,7 +5,8 @@
   "main": "script.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server.js"
+    "start": "node server.js",
+    "build": "webpack"
   },
   "repository": {
     "type": "git",
@@ -17,5 +18,15 @@
   "bugs": {
     "url": "https://github.com/natakoltunova/js2/issues"
   },
-  "homepage": "https://github.com/natakoltunova/js2#readme"
+  "homepage": "https://github.com/natakoltunova/js2#readme",
+  "devDependencies": {
+    "webpack": "^5.8.0",
+    "webpack-cli": "^4.2.0"
+  },
+  "dependencies": {
+    "@babel/core": "^7.12.9",
+    "@babel/plugin-proposal-class-properties": "^7.12.1",
+    "babel": "^6.23.0",
+    "babel-loader": "^8.2.2"
+  }
 }
