+++ public/index.html
@@ -25,7 +25,6 @@
       </p>
     </main>
     <footer></footer>
-    <script src="button.js"></script>
-    <script src="shop.js"></script>
+    <script src="./js/shop.js"></script>
   </body>
 </html>
