+++ public/js/shop.js
@@ -1,0 +1,1 @@
+(()=>{"use strict";class e{constructor(e=[]){var t,n;n=[],(t="items")in this?Object.defineProperty(this,t,{value:n,enumerable:!0,configurable:!0,writable:!0}):this[t]=n,this.item=[]}add(e){const t=this.items.find((t=>t.name===e.name));new Promise((n=>{t?t.counter++:this.items.push(e),n()})).then(this.render.bind(this))}render(){this.items.forEach((e=>{e.render()}))}}function t(e,t,n){return t in e?Object.defineProperty(e,t,{value:n,enumerable:!0,configurable:!0,writable:!0}):e[t]=n,e}class n{constructor(e,n){t(this,"_text",""),t(this,"_onClickClb",null),this._text=e,this._onClickClb=n}get text(){return this._text}set text(e){this._text=e}onBtnClick(){console.log("Clicked!"),"function"==typeof this._onClickClb&&this._onClickClb()}getMainTemplate(){const e=document.createElement("div");return e.classList.add("btn"),e}getTemplate(){const e=this.getMainTemplate();return e.innerHTML=this.text,e.addEventListener("click",(()=>{this.onBtnClick()})),e}}class s extends e{constructor(){super(),this.init()}init(){const e=document.createElement("div");e.classList.add("cart");const t=document.createElement("div");t.classList.add("cart_list"),e.appendChild(t);const s=new n("Корзина",(()=>{t.classList.toggle("shown")}));e.appendChild(s.getTemplate());const i=document.querySelector("header");i&&i.appendChild(e),this.render()}render(){const e=document.querySelector(".cart_list");e&&(e.innerHTML="",this.items.length?this.items.forEach((t=>{t.render(e)})):e.innerHTML="Нет товаров в корзине!")}}new s;class i{constructor(){this.render()}render(){const e=document.querySelector("footer");if(e){e.innerHTML='\n        <form class="contact" action="">\n          <fieldset>\n            <legend>Форма обратной связи</legend>\n            <p class="contact_line"><label for="name">Имя <input class="contact_field" placeholder=" Your name" name="username" id="name"></p>\n            <p class="contact_line"><label for="phone">Телефон </label><input class="contact_field" placeholder=" +7(000)000-0000" name="phone" id="phone"></p>\n            <p class="contact_line"><label for="email">E-mail </label><input class="contact_field" placeholder=" valid email" name="email" id="email"></p>\n          </fieldset>\n        <button type="submit">Отправить</button>\n        <button type="reset">Сбросить</button>\n      </form>\n        ';const t=document.querySelectorAll("input"),n={username:/^[A-Za-z]+$/,phone:/^\+7{1}\(\d{3}\)\d]{3}-\d{4}$/,email:/^([a-z\d\.-]+)@([a-z]+)\.([a-z]{2})$/};t.forEach((e=>{e.addEventListener("keyup",(e=>{var t,s;t=e.target,s=n[e.target.attributes.name.value],console.log(s.test(t.value)),s.test(t.value)?t.className="valid":t.className="invalid"}))}))}}}new i;class a extends c{constructor(e){super(e)}render(e){if(e){const t=document.createElement("div");t.classList.add("cart__good"),t.innerHTML=`${this.name}  (${this.price})  ${this.counter} шт. `,e.appendChild(t);const s=new n("убрать");t.appendChild(s.getTemplate())}}}function r(e,t,n){return t in e?Object.defineProperty(e,t,{value:n,enumerable:!0,configurable:!0,writable:!0}):e[t]=n,e}class c{constructor({name:e,price:t},n){r(this,"name",""),r(this,"price",0),r(this,"counter",1),r(this,"_cartInstance",null),this.name=e,this.price=t,this._cartInstance=n}render(e){if(e){const t=document.createElement("div");t.classList.add("good-item"),t.innerHTML=`\n        <div class="img">\n          <img src="./img/bossparts.jpg" />\n        </div>\n        <div class="meta">\n          <div class="meta__row">\n            <span class="key">Товар:</span>\n            <span class="value">${this.name}</span>\n          </div>\n          <div class="meta__row">\n            <span class="key">Цена:</span>\n            <span class="value">${this.price}</span>\n          </div>\n          <div class="btn_holder"></div>\n        </div>\n      `,e.appendChild(t);const s=new n("В корзину",(()=>{this._cartInstance.add(new a(this))}));t.querySelector(".btn_holder").appendChild(s.getTemplate())}}}function o(e,t,n){return t in e?Object.defineProperty(e,t,{value:n,enumerable:!0,configurable:!0,writable:!0}):e[t]=n,e}class l extends e{constructor(e){super(),o(this,"_cartInstance",null),o(this,"_pageCounter",1),this._cartInstance=e,this.initShowMoreBtn(),this.fetchGoods().then((()=>{this.render()}))}initShowMoreBtn(){document.querySelector(".showmore").addEventListener("click",(()=>{this.fetchGoods().then((()=>{this.render()}))}))}hideShowMoreBtn(){document.querySelector(".showmore").remove()}fetchGoods(){return fetch(`./database/page${this._pageCounter}.json`).then((e=>e.json())).then((e=>{this._pageCounter++,this.items.push(...e.data.map((e=>new c(e,this._cartInstance))))})).catch((e=>{this.hideShowMoreBtn(),console.log(e)}))}render(){const e=document.querySelector(".goods-list");e&&(e.innerHTML="",this.items.forEach((t=>{t.render(e)})))}}new l(CartInstance),new l(new s),new i})();
