+++ src/.babelrc
@@ -1,0 +1,3 @@
+{
+  "plugins": ["@babel/plugin-proposal-class-properties"]
+}
